+++ SSIS-WebAPI/SSIS.DataService/SQL+/Queries/SSISDB/ExeuctionMessages.sql
@@ -9,7 +9,7 @@
     DECLARE
     
     --+Required
-    @ExecutionId int = 11,
+    @ExecutionId bigint = 11,
 
     --+Default=false
     --+Comment=Include only basic logging messages. Excludes validation messages.
